services/recurring-task-service/requirements.txt
# T005 - Recurring Task Service Python dependencies
# Phase V Todo Chatbot - Recurring Task Service

# Web framework for Dapr subscriptions
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Data validation
pydantic==2.6.1

# HTTP client for Dapr
httpx==0.27.0

# Date handling for recurrence calculation
python-dateutil==2.8.2

# Logging
structlog==24.1.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
